+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Year-aware helper tasks
 - Helper task validation
+- Consistent time zone handling for datetime properties
 
 ### Changed
 
@@ -26,6 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - OracleDB adapter 2.0.1 & migrate to full async usage
 - Upgrade to Python 3.11
 - Update REST API endpoints to /v1/
+
+TODO tests
+TODO other todos
 
 ## [0.4.0] - 2023-06-26